fix the pitches on OPNA

+++ src/chips/gens_opn2.cpp
@@ -37,7 +37,7 @@ GensOPN2::~GensOPN2()
 void GensOPN2::setRate(uint32_t rate, uint32_t clock)
 {
     OPNChipBaseBufferedT::setRate(rate, clock);
-    uint32_t chipRate = isRunningAtPcmRate() ? rate : static_cast<uint32_t>(nativeRate);
+    uint32_t chipRate = isRunningAtPcmRate() ? rate : nativeRate();
     chip->set_rate(chipRate, clock);  // implies reset()
 }
